+++ c-ares/src/query.rs
@@ -21,7 +21,7 @@ pub(crate) unsafe extern "C" fn raw_query_callback<F>(
         Err(Error::from(status))
     };
     let handler = unsafe { Box::from_raw(arg.cast::<F>()) };
-    panic::catch(|| handler(result));
+    panic::abort_on_panic(|| handler(result));
 }
 
 pub(crate) unsafe extern "C" fn query_callback<R, F>(